Update index.html
Updated by Grok after I assessed previous commit. Then I prompted: I've deployed the latest code. It looks better but there a couple of things that broke. 1. the "1 Verify Website" indicator is gray instead of blue on stage 1. and 2. the "Check Website Compatibility" button does not click through to the next step so I can see how the rest of the code executes.